se termino el microservicios de catalogo
se completo el microservicio de catalogo respetando los codigos de error y agregado el DTOestandar con el global exception

+++ microservicio catalogo/src/main/java/com/example/catalogo/dto/SeriesDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SeriesDTO {
     @NotBlank(message= "El Id del ejercicio no puede estar vacio")
-    private int EjercicioId;
+    private Long EjercicioId;
 
     @Min(value = 0, message= "El eso debe ser positivo")
     private double peso;
